+++ lib/overcommit/hook/pre_commit/hard_tabs.rb
@@ -2,10 +2,8 @@ module Overcommit::Hook::PreCommit
   # Checks for hard tabs in files.
   class HardTabs < Base
     def run
-      paths = applicable_files.join(' ')
-
       # Catches hard tabs
-      result = command("grep -IHn \"\t\" #{paths}")
+      result = command(%w[grep -IHn] + ["\t"] + applicable_files)
       unless result.stdout.empty?
         return :bad, "Hard tabs detected:\n#{result.stdout}"
       end