+++ lib/overcommit/hook/pre_commit/rubocop.rb
@@ -6,11 +6,13 @@ module Overcommit::Hook::PreCommit
         return :warn, 'Rubocop not installed -- run `gem install rubocop`'
       end
 
-      result = command("rubocop --format=emacs #{applicable_files.join(' ')} 2>&1")
+      result = command(%w[rubocop --format=emacs] + applicable_files)
       return :good if result.success?
 
+      output = result.stdout + result.stderr
+
       # Keep lines from the output for files that we actually modified
-      error_lines, warning_lines = result.stdout.split("\n").partition do |output_line|
+      error_lines, warning_lines = output.split("\n").partition do |output_line|
         if match = output_line.match(/^([^:]+):(\d+)/)
           file = match[1]
           line = match[2]