+++ spec/overcommit/hook/pre_commit/bundle_check_spec.rb
@@ -38,7 +38,7 @@ describe Overcommit::Hook::PreCommit::BundleCheck do
     before do
       result.stub(:success? => success, :stdout => 'Bundler error message')
       subject.stub(:command).and_call_original
-      subject.stub(:command).with('bundle check').and_return(result)
+      subject.stub(:command).with(%w[bundle check]).and_return(result)
     end
 
     context 'and bundle check exits unsuccessfully' do